Fixed install issues
Replaced all references to scipy's toimage function with a copy of it in a separate file. This prevents the user from having to deal with installing deprecated scipy stuff. Also added the complete list of imagenet classes to a separate TXT file so you don't need to look them up and try to figure out which list is correct.